% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gassplit.R
\name{essp.gasgroups}
\alias{essp.gasgroups}
\title{Natural gas prime-mover groups}
\usage{
essp.gasgroups()
}
\value{
A tibble of \code{group}, \code{codes}, and \code{role}.
}
\description{
The mapping from EIA prime-mover codes to the plant types a demand curve
distinguishes.
}
\examples{
essp.gasgroups()

}
