% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{essp_facet_ids}
\alias{essp_facet_ids}
\title{Facet ids available on a route}
\usage{
essp_facet_ids(route)
}
\arguments{
\item{route}{An EIA route path.}
}
\value{
A character vector of facet ids, empty if the route has none.
}
\description{
Facet ids available on a route
}
\keyword{internal}
