% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{essp_frequencies}
\alias{essp_frequencies}
\title{Frequencies available on a route}
\usage{
essp_frequencies(route)
}
\arguments{
\item{route}{An EIA route path.}
}
\value{
A character vector of frequency ids, empty if the route declares none.
}
\description{
Frequencies available on a route
}
\keyword{internal}
